pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-story-maker"
version = "0.1.0"
description = "Interactive Story Generator with AI"
requires-python = ">=3.13"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.15",
    "psycopg[binary]>=3.2",
    "celery>=5.4",
    "django-celery-results>=2.5",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "gunicorn>=23.0",
    "whitenoise>=6.7",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "mypy>=1.13",
    "pre-commit>=4.0",
    "django-stubs>=5.1",
    "djangorestframework-stubs>=3.15",
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=4.1",
    "pytest-playwright>=0.5",
    "factory-boy>=3.3",
]

[tool.setuptools.packages.find]
include = ["apps*", "config*"]

[tool.ruff]
target-version = "py313"
line-length = 88
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
    "migrations",
    "staticfiles",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "DJ",     # flake8-django
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "DJ001",  # null=True on string fields (intentional for selected_choice)
]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.mypy]
python_version = "3.13"
plugins = ["mypy_django_plugin.main"]
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = [
    "migrations/",
    "venv/",
    ".venv/",
]

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "django_celery_results.*",
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["test_*.py", "*_test.py"]
addopts = [
    "--strict-markers",
    "-ra",
]
testpaths = ["tests", "apps"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "e2e: marks tests as end-to-end browser tests (deselect with '-m \"not e2e\"')",
]

[tool.coverage.run]
source = ["apps", "common"]
branch = true
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/__pycache__/*",
    "*/admin.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
fail_under = 80
show_missing = true

[tool.coverage.html]
directory = "htmlcov"
